move BlockAdvancement to app testutils
